+++ crates/rtlola-interpreter-macros/src/value_factory/struct_factory.rs
@@ -114,7 +114,7 @@ pub(crate) fn expand_unnamed_struct(input: &DeriveInput) -> TokenStream {
     let type_args: Vec<_> = generics.params.iter().collect();
 
     quote! {
-        impl #impl_generics rtlola_interpreter::input::AssociatedFactory for #name #ty_generics #where_clause {
+        impl #impl_generics rtlola_interpreter::input::AssociatedEventFactory for #name #ty_generics #where_clause {
             type Factory = #factory #ty_generics;
         }
 
@@ -149,7 +149,7 @@ pub(crate) fn expand_unit_struct(input: &DeriveInput) -> TokenStream {
     let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
 
     quote! {
-        impl #impl_generics rtlola_interpreter::input::AssociatedFactory for #name #ty_generics #where_clause {
+        impl #impl_generics rtlola_interpreter::input::AssociatedEventFactory for #name #ty_generics #where_clause {
             type Factory = rtlola_interpreter::input::EmptyFactory<Self>;
         }
     }