+++ crates/rtlola-io-plugins/examples/tutorial.rs
@@ -1,7 +1,7 @@
 use std::error::Error;
 use std::net::{TcpStream, UdpSocket};
 
-use rtlola_interpreter::input::AssociatedFactory;
+use rtlola_interpreter::input::AssociatedEventFactory;
 use rtlola_interpreter::monitor::TriggerMessages;
 use rtlola_interpreter::time::{AbsoluteFloat, RealTime};
 use rtlola_interpreter::ConfigBuilder;
@@ -46,7 +46,7 @@ fn main() -> Result<(), Box<dyn Error + 'static>> {
     let mut monitor = ConfigBuilder::new()
         .spec_str(SPEC)
         .online()
-        .with_event_factory::<<ExampleInputs as AssociatedFactory>::Factory>()
+        .with_event_factory::<<ExampleInputs as AssociatedEventFactory>::Factory>()
         .with_verdict::<TriggerMessages>()
         .output_time::<AbsoluteFloat>()
         .monitor()?;