+++ crates/rtlola-interpreter/CHANGELOG.md
@@ -8,9 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `FromValues` trait and `StructVerdictFactory` to easily construct custom datatypes from monitor verdicts.
+- `AssociatedVerdictFactory` that links a type to the respective `VerdictFactory`.
 
 ### Changed
 - The trigger field in `IncrementalVerdict` now contains `TriggerReferences`, not `OutputReferences`.
+- Renamed `AssociatedFactory` to `AssociatedEventFactory`.
 
 ### Fixed
 - Fix bug where `Incremental` verdict representation included trigger twice.